audiovisual_jobs/requirements.txt
python-dotenv>=1.0.0
supabase>=2.0.0
playwright>=1.40.0
telethon>=1.30.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
python-dateutil>=2.8.0
google-search-results>=2.4.0
tqdm>=4.66.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
